pyhq refactoring
