+++ crates/pyhq/python/hyperqueue/utils/package.py
@@ -1,9 +1,3 @@
-def pluralize(text: str, count: int) -> str:
-    if count == 1:
-        return text
-    return f"{text}s"
-
-
 class MissingPackageException(BaseException):
     def __init__(self, package: str):
         self.package = package
