+++ crates/pyhq/python/hyperqueue/utils/string.py
@@ -1,0 +1,4 @@
+def pluralize(text: str, count: int) -> str:
+    if count == 1:
+        return text
+    return f"{text}s"
